% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NestedOrNotFunction.R
\name{NestedOrNot}
\alias{NestedOrNot}
\title{NestedOrNot Function}
\usage{
NestedOrNot(Samplecov1, FactorModel1, Size1, Samplecov2, FactorModel2, Size2)
}
\arguments{
\item{"Samplecov1", }{"Samplecov2", "FactorModel1", "FactorModel2", "Size1" and "Size2" needed for this function.}
}
\description{
This function allows you to know two SEM models are nested in terms of tetrads or not.
}
\examples{
NestedOrNot()
}
\keyword{Analysis}
\keyword{Confirmatory}
\keyword{Nested}
\keyword{Tetrad}
\keyword{in}
\keyword{not}
\keyword{or}
