% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NestedOrNotFunction.R
\name{NestedOrNot}
\alias{NestedOrNot}
\title{NestedOrNot Function}
\usage{
NestedOrNot(Samplecov1, FactorModel1, Size1, Samplecov2, FactorModel2, Size2)
}
\arguments{
\item{"Data1", }{An optional data frame containing the observed variables used in the first model.}

\item{"Data2", }{An optional data frame containing the observed variables used in the second model.}

\item{"Model1"}{A description of the user-specified first model. The model is described using the lavaan model synta(Rosseel, 2012).}

\item{"Model2"}{A description of the user-specified second model. The model is described using the lavaan model synta(Rosseel, 2012).}

\item{"Size1"}{Size of the first data sample.}

\item{"Size2"}{Size of the second data sample.}
}
\value{
"Model1COV",    Implied variance-covariance matrix from the first model.

"Model2COV",    Implied variance-covariance matrix from the second model.

"NRVT1",     Nonredundant vanishing tetrad setting implied by the first model.

"NRVT2",     Nonredundant vanishing tetrad setting implied by the second model.

"NRVT_Num1",     Nonredundant vanishing tetrad number for the first model.

"NRVT_Num1",     Nonredundant vanishing tetrad number for the second model.

"NEST",     Two models are nested or not nested in terms of tetrads.
}
\description{
"NestedOrNot " is used to know two models are nested in terms of tetrads or not (Bollen, 1993).
}
\examples{
## Input sample variance-covariance matrix "(Sampledata) ".
data<-c(2.18, 0.63, 0.50, 0.22, 0.39, 0.42, 0.63, 1.44, 0.50, 0.12, 0.11, 0.15, 0.50, 0.50, 1.71, 0.87, 0.59, 0.48, 0.22, 0.12, 0.87, 1.46, 0.54, 0.42, 
0.39, 0.11, 0.59, 0.54, 1.12, 0.23, 0.42, 0.15, 0.48, 0.42, 0.23, 1.42)
Sampledata <- matrix(data, nrow = 6, ncol = 6, byrow = TRUE)
colnames(Sampledata) <- c('x1','x2','x3','x4','x5','x6')

## Sample size
Size<-100

## Input model
FactorModel1<-'
Y1 =~ x1 + x2 + x3
Y2 =~ x4 + x5 + x6
Y2 ~ Y1
x3 ~ x4
'

FactorModel2<-'
Y1 =~ x1 + x2 + x3
Y2 =~ x4 + x5 + x6
Y2 ~ Y1
x4 ~~ x3
'

## Use "NestedOrNot " to run Confirmatory Tetrad Analysis
NestedOrNot(Sampledata,FactorModel1,Size,Sampledata,FactorModel2,Size)
}
\references{
Bollen, K. A., & Ting, K. F. (1993). Confirmatory tetrad analysis. Sociological methodology, 147-175.

Rosseel, Y. (2012). lavaan: An R package for structural equation modeling. Journal of Statistical Software, 48(2), 1-36.
}
\keyword{Analysis}
\keyword{Confirmatory}
\keyword{Nested}
\keyword{Tetrad}
\keyword{in}
\keyword{not}
\keyword{or}
