% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tetrad Function.R
\name{TetradAnalysisNoRandom}
\alias{TetradAnalysisNoRandom}
\title{Confirmatory Tetrad Analysis Function}
\usage{
TetradAnalysisNoRandom(SampleData, ModelFormula, SampleSize)
}
\arguments{
\item{"Samplecov", An optional data frame containing the observed variables used in the model. }
\item{"ModelFormula", A description of the user-specified model. Typically, the model is described using the lavaan model syntax. }
\item{"SampleSize", Size of sample data.}
}
\description{
This function is used to run Confirmatory Tetrad Analysis(from Bollen 1993).
}
\Value{
\item{"T", Test statistic for multivariate test in confirmatory tetrad analysis}
\item{"MultiPvalue", P-value for multivariate test in confirmatory tetrad analysis}
\item{"NRVT", Nonredundant vanishing tetrad setting for inputted model}
\item{"NRVT", Nonredundant vanishing tetrad setting for inputted model}  
}
\examples{
TetradAnalysisNoRandom()
}
\keyword{Analysis}
\keyword{Confirmatory}
\keyword{Tetrad}
