% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tetrad Function.R
\name{TetradAnalysisNoRandom}
\alias{TetradAnalysisNoRandom}
\title{Confirmatory Tetrad Analysis Function}
\usage{
TetradAnalysisNoRandom(Samplecov, FactorModel, Size)
}
\arguments{
\item{"Samplecov", }{"FactorModel", and "Size"}
}
\description{
This function allows you to run Confirmatory Tetrad Analysis from Bollen 1993.
}
\examples{
TetradAnalysisNoRandom()
}
\keyword{Analysis}
\keyword{Confirmatory}
\keyword{Tetrad}
